+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
-## Unreleased
+## [1.0.1] - 2023-03-28
 
 ### Changed
 
@@ -189,6 +189,7 @@ The two most important changes in usage are:
 [0.5.0]: https://github.com/dogmatiq/ferrite/releases/tag/v0.5.0
 [0.6.0]: https://github.com/dogmatiq/ferrite/releases/tag/v0.6.0
 [1.0.0]: https://github.com/dogmatiq/ferrite/releases/tag/v1.0.0
+[1.0.1]: https://github.com/dogmatiq/ferrite/releases/tag/v1.0.1
 
 <!-- version template
 ## [0.0.1] - YYYY-MM-DD